added create_booth template, create_event template
create_booth - admins are now able to add new organizations to the
system in order to update passport. The show view currently display all
the booths in the system.

create_event - admins are able to create "events" in order to display
information to users. checkboxes currently do not send any information.